Fixes #3: Evaluate check hooks
This sets them up as far as I can see. They have an entry in the event
in uchiwa.

There is NO evidence that the associated command is invoked. There is NO
output from that command, and side effects of the command are not
visible.

I guess this is a failed evaluation.